move helper files to lib directory
