+++ src/api/transform/lib/decodeRaw.ts
@@ -1,5 +1,5 @@
-import { Text } from '../../core/domain/domains.ts'
-import { Code } from '../../core/code/code.ts'
+import { Text } from '../../../core/domain/domains.ts'
+import { Code } from '../../../core/code/code.ts'
 
 // takes a transitional object and returns a new object with the same code and a new raw Buffer with padding intact
 export const decodeRaw = (transitional: {
