+++ src/api/transform/lib/removePadding.ts
@@ -1,5 +1,5 @@
-import { Raw } from '../../core/domain/domains.ts'
-import { Code } from '../../core/code/code.ts'
+import { Raw } from '../../../core/domain/domains.ts'
+import { Code } from '../../../core/code/code.ts'
 
 // takes a transitional object removes any padding from 'raw' and returns a new Raw object
 export const removePadding = (transitional: {
